Share tel/WhatsApp phone links via a reusable component
Extract the attendee-detail "phone + tel + WhatsApp links" markup into a
shared PhoneLinks component (alongside MapsLinks) and use it on the agent
run sheet, so deliveries get the same tel: and WhatsApp links instead of a
tel-only link. attendee-detail now uses the shared component too, removing
the duplicate.
